backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-socketio==5.10.0
aiohttp==3.9.1
httpx==0.26.0
openai==1.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
